update- reorganize user classes between packages
