+++ src/main/java/demo/codeexample/config/DataInitializer.java
@@ -1,7 +1,7 @@
 package demo.codeexample.config;
 
 import demo.codeexample.user.doman.Role;
-import demo.codeexample.user.User;
+import demo.codeexample.user.doman.User;
 import demo.codeexample.user.doman.UserRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.security.crypto.password.PasswordEncoder;
