+++ src/main/java/demo/codeexample/security/OAuth2LoginSuccessHandler.java
@@ -1,7 +1,7 @@
 package demo.codeexample.security;
 
 import demo.codeexample.user.doman.Role;
-import demo.codeexample.user.*;
+import demo.codeexample.user.doman.User;
 import demo.codeexample.user.doman.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
